# gateway-api: avoid deep copy when converting http routes