**Please provide a description of this PR:**
Shallow-copy `spec.Rules` instead of deep-copying the whole object
Trying to make further optimization based on https://github.com/istio/istio/pull/37139#discussion_r799075531