@istio-testing 2022-05-23
@dddddai: The following tests **failed**, say `/retest` to rerun all failed tests or `/retest-required` to rerun all mandatory failed tests:

Test name | Commit | Details | Required | Rerun command
--- | --- | --- | --- | ---
lint_istio | caa1f22fdc6282e58dc6b6f1200ec116434eec92 | [link](https://prow.istio.io/view/gs/istio-prow/pr-logs/pull/istio_istio/39085/lint_istio/1528728594505797632) | true | `/test lint_istio`
gencheck_istio | caa1f22fdc6282e58dc6b6f1200ec116434eec92 | [link](https://prow.istio.io/view/gs/istio-prow/pr-logs/pull/istio_istio/39085/gencheck_istio/1528728595340464128) | true | `/test gencheck_istio`

<details>

Instructions for interacting with me using PR comments are available [here](https://git.k8s.io/community/contributors/guide/pull-requests.md).  If you have questions or suggestions related to my behavior, please file an issue against the [kubernetes/test-infra](https://github.com/kubernetes/test-infra/issues/new?title=Prow%20issue:) repository. I understand the commands that are listed [here](https://go.k8s.io/bot-commands).
</details>
<!-- test report -->