@dddddai 2022-06-21
Benchmark results on my laptop:
```
name                        old time/op    new time/op    delta
BuildHTTPVirtualServices-4     104µs ± 1%     101µs ± 3%  -2.75%  (p=0.003 n=6+11)

name                        old alloc/op   new alloc/op   delta
BuildHTTPVirtualServices-4    28.1kB ± 0%    26.6kB ± 0%  -5.41%  (p=0.000 n=7+12)

name                        old allocs/op  new allocs/op  delta
BuildHTTPVirtualServices-4       485 ± 0%       455 ± 0%  -6.19%  (p=0.000 n=7+12)
```

The delta can be greater if HTTPRoute spec is larger
@howardjohn Could you please take a look?